backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
pymongo==4.8.0
dnspython==2.6.1